Refactored test files
Renamed most function called to `process` for better consistency and
readability across modules.
