+++ lib/ecto/fixtures/conditioners/override.ex
@@ -1,5 +1,5 @@
 defmodule EctoFixtures.Conditioners.Override do
-  def override(data, [source: source, override: %{}=override_data, reverse: reverse?]=opts) do
+  def process(data, [source: source, override: %{}=override_data, reverse: reverse?]=opts) do
     Enum.reduce override_data, data, fn({table_name, rows}, data) ->
       case get_in(data, [source, table_name]) do
         nil -> data
@@ -7,21 +7,21 @@ defmodule EctoFixtures.Conditioners.Override do
           result = case get_in(data, [source, table_name, :rows, row_name]) do
             nil -> data
             _ ->
-              put_in(data, [source, table_name, :rows, row_name, :data], override_merge(get_in(data, [source, table_name, :rows, row_name, :data]), columns, reverse?))
+              put_in(data, [source, table_name, :rows, row_name, :data], merge(get_in(data, [source, table_name, :rows, row_name, :data]), columns, reverse?))
           end
         end
       end
     end
   end
-  def override(data, [source: source, override: override_data]) when is_map(override_data), do:
-    override(data, [source: source, override: override_data, reverse: false])
-  def override(data, _opts), do: data
+  def process(data, [source: source, override: override_data]) when is_map(override_data), do:
+    process(data, [source: source, override: override_data, reverse: false])
+  def process(data, _opts), do: data
 
-  defp override_merge(left, right, false) do
+  defp merge(left, right, false) do
     Map.merge(left, right)
   end
 
-  defp override_merge(left, right, true) do
+  defp merge(left, right, true) do
     Map.merge(right, left)
   end
 end
